+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - 2026-01-05
+
+- Added `until(() => bool)` for behavior based on a boolean condition
+**please see examples to avoid bugs or loops**
+---
+
 ## [0.3.0] - 2026-01-05
 
 - Added `onCall(n)` for behavior on call N